src/choice/game/requirements.txt
flask
pyopenssl